#55: refactoring as preparation for better testing
