+++ src/main/resources/com/microsoft/jenkins/containeragents/aci/AciContainerTemplate/help-usePrivateIpAddress.html
@@ -1,3 +1,5 @@
 <div>
-    Select if the container instance will use a public ip or a private ip address.
+    Select if the container instance will use a private ip address.
+
+    If it use a private ip address, then information about the vnet and the subnet are mandatory.
 </div>
